feat: 3단계 - 질문 삭제하기 리팩터링

+++ src/test/java/qna/domain/UserTest.java
@@ -4,10 +4,10 @@ public class UserTest {
     public static final User JAVAJIGI = new User(1L, "javajigi", "password", "name", "javajigi@slipp.net");
     public static final User SANJIGI = new User(2L, "sanjigi", "password", "name", "sanjigi@slipp.net");
 
-    private static int id = 1;
+    private static long id = 1;
 
     public static User newInstance() {
-        int no = id++;
+        long no = id++;
         return User.builder()
                 .userId("test" + no)
                 .password("pass1234")
